Initial import from MediaWiki core
The code is a straight copy from MediaWiki core, except that I moved up
the file comments, removed @since tags, and the deprecated class_alias()
calls.

The README needs to be filled out still, but that can be done in
follow-ups by someone who is more familiar with how the code should be
instantiated/used.

Bug: T211608
Change-Id: I89a91644dbb00625c36421f71b160dc5a8b5a8ac