Improve performance of very hot methods
While this is not a bottleneck it's still notable because it's
called literally millions of times. The function call overhead is
quite large in PHP. And we don't really need the extra functionality
provided by the peekService() method: it checks if the service name
is valid. But it's not critical or even impossible to be invalid in
these contexts.

In my local tests the total runtime of esp. resetService() is
dramatically improved: about 15 times faster.

Change-Id: I5260ea7934b9830b8fa229af2ed0d507c5f75da8